Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-   float x;
-   scanf("%.4f",&x);
-   printf("You entered: %.4f",x);
+   double x;
+   scanf("%.4lf",&x);
+   printf("You entered: %.4lf",x);
    return 0;
 }
